pyproject.toml
[project]
name = "project-fine-tuning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bitsandbytes>=0.46.1",
    "dataset>=1.6.2",
    "peft>=0.18.1",
    "trl>=1.0.0",
]

[dependency-groups]
dev = [
    "ruff>=0.15.9",
]
